requirements.txt
langchain
langchain-core
langchain-openai
langchain-experimental
python-dotenv
